requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==24.2.0
automaton==3.2.0
autopage==0.5.2
babel==2.16.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cliff==4.8.0
cloudevents==1.9.0
cmd2==2.5.7
comm==0.2.2
cryptography==44.0.0
dataclasses-json==0.6.7
debtcollector==3.0.0
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
dogpile.cache==1.3.3
email_validator==2.2.0
eventlet==0.38.0
executing==2.1.0
fastapi==0.115.4
fastapi-cli==0.0.5
fasteners==0.19
fastjsonschema==2.21.1
firebase-admin==6.6.0
firebase-functions==0.1.2
Flask==3.1.0
Flask-Cors==5.0.0
fqdn==1.5.1
frozenlist==1.5.0
functions-framework==3.8.2
futurist==3.0.0
gnureadline==8.2.13
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-events==0.13.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
iso8601==2.1.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keystoneauth1==5.9.1
keystonemiddleware==10.8.0
kombu==5.4.2
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-openai==0.2.10
langchain-text-splitters==0.3.2
langgraph==0.2.53
langgraph-checkpoint==2.0.7
langgraph-sdk==0.1.40
langsmith==0.1.147
legacy-cgi==2.6.1
logutils==0.3.5
magnum==19.0.0
Mako==1.3.7
mangum==0.19.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==1.3.0
netifaces==0.11.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
# numpy==2.1.3
openai==1.56.0
# Perplexity uses OpenAI-compatible API, so no separate package needed
openstacksdk==4.1.0
orjson==3.10.12
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==3.2.0
oslo.cache==3.9.0
oslo.concurrency==6.2.0
oslo.config==9.7.0
oslo.context==5.7.0
oslo.db==17.0.0
oslo.i18n==6.5.0
oslo.log==6.2.0
oslo.messaging==15.0.0
oslo.metrics==0.10.0
oslo.middleware==6.3.0
oslo.policy==4.5.0
oslo.reports==3.5.0
oslo.serialization==5.6.0
oslo.service==3.6.0
oslo.upgradecheck==2.4.0
oslo.utils==7.4.0
oslo.versionedobjects==3.5.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
Paste==3.10.1
PasteDeploy==3.1.0
pbr==6.1.0
pecan==1.5.1
pexpect==4.9.0
platformdirs==4.3.6
prettytable==3.12.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.0
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycadf==4.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydot==3.0.3
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyparsing==3.2.0
pyperclip==1.9.0
python-barbicanclient==7.0.0
python-cinderclient==9.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-glanceclient==4.7.0
python-heatclient==4.0.0
python-json-logger==2.0.7
python-keystoneclient==5.5.0
python-multipart==0.0.17
python-neutronclient==11.3.1
python-novaclient==18.7.0
python-octaviaclient==3.9.0
python-openstackclient==7.2.1
python-swiftclient==4.6.0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
repoze.lru==0.7
requests==2.32.3
requests-toolbelt==1.0.0
requestsexceptions==1.4.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.9.4
Routes==2.5.1
rpds-py==0.22.0
rsa==4.9
Send2Trash==1.8.3
setuptools==75.6.0
shellingham==1.5.4
simplegeneric==0.8.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.2
statsd==4.0.1
stevedore==5.4.0
taskflow==5.10.0
tavily-python==0.5.0
tenacity==9.0.0
terminado==0.18.1
testresources==2.0.1
testscenarios==0.5.0
testtools==2.7.2
tiktoken==0.8.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
twilio==9.3.7
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
vine==5.1.0
warlock==2.0.1
watchdog==6.0.0
watchfiles==0.24.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
WebOb==1.8.9
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wikipedia==1.4.0
wrapt==1.17.0
WSME==0.12.1
yappi==1.6.10
yarl==1.18.3
zipp==3.21.0
